added Blueprints for mscolab
